requirements.txt
langchain==0.3.26
langchain-community==0.3.26
flask==3.1.1
pypdf==5.6.1
python-dotenv==1.1.0
langchain-pinecone
langchain-groq
gunicorn
pinecone
fastembed
